+++ src/main/java/lotto/Application.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.domain.LottoGame;
+import lotto.controller.LottoGame;
 
 public class Application {
